+++ libs/shared/src/lib/layouts/index.ts
@@ -1,2 +1,3 @@
 export { LoginLayoutComponent } from './login-layout/login-layout.component';
 export { MainLayoutComponent } from './main-layout/main-layout.component';
+export * from './top-bar';