+++ libs/shared/src/lib/models/index.ts
@@ -1,4 +1,5 @@
 export * from './auth-response';
 export * from './current-user-info';
+export * from './notification';
 export * from './operations-history';
 export * from './services-history';