+++ libs/shared/src/lib/services/index.ts
@@ -1,3 +1,4 @@
 export { AuthenticationService } from './authentication.service';
+export { NotificationsService } from './notifications.service';
 export { OperationsHistoryService } from './operations-history.service';
 export { ServicesHistoryService } from './services-history.service';